feat(board): 添加PC通信功能并优化硬件配置
- 添加PC串口通信驱动程序和相关头文件
- 集成UART7作为PC通信接口并配置GPIO引脚PE7和PE8
- 在ioc配置文件中更新STM32Cube项目设置和引脚分配
- 添加Chassis_data_ID_3用于底盘数据传输
- 修复HWT906角度数据转换精度问题
- 调整VT03遥控器模式开关逻辑和射击控制参数
- 更新CAN数据帧发送机制以支持PC数据传输
- 修改调试服务器ST-LINK配置和构建选项字符串

+++ User_Drives/user_HWT906.h
@@ -28,7 +28,7 @@ typedef struct {
 typedef struct {
     short angle_x;  // x轴角度
     short angle_y;  // y轴角度
-    short angle_z;  // z轴角度
+    float angle_z;  // z轴角度
     short version;  // 版本号
 }ANGLE;
